[evals] Break trace label import cycle
