+++ lib/levanter/src/levanter/data/text/__init__.py
@@ -30,6 +30,7 @@ from .examples import (
     GrugLmExample,
     LabeledLmExample,
     LOSS_IGNORE_LABEL,
+    LossLabelSpec,
     LossLabelSpan,
     grug_attention_mask_from_named,
     grug_lm_example_from_named,
@@ -88,6 +89,7 @@ __all__ = [
     "GrugLmExample",
     "LabeledLmExample",
     "LOSS_IGNORE_LABEL",
+    "LossLabelSpec",
     "LossLabelSpan",
     "grug_attention_mask_from_named",
     "grug_lm_example_from_named",
